Added new test cases!
